requirements.txt
fastapi
hypercorn
catboost
pandas
fasttext-wheel
nltk